+++ frontend/src/components/TableReport.vue
@@ -13,9 +13,9 @@
                     </template>
                     <v-list>
                         <v-list-item-group>
-                            <v-list-item v-for="(item, index) in exportTypes" :key="index">
+                            <v-list-item v-for="(item, index) in exportTypes" :key="index" @click="goExport(item)">
                                 <v-list-item-content>
-                                    <v-list-item-title @click="goExport(item)">{{item}}</v-list-item-title>
+                                    <v-list-item-title>{{item}}</v-list-item-title>
                                 </v-list-item-content>
                             </v-list-item>
                         </v-list-item-group>
@@ -33,6 +33,7 @@
         <v-row>
             <v-col cols=12>
                 <v-data-table
+                    id="reportTable"
                     dense
                     :headers="headers"
                     :items="items"
@@ -48,6 +49,10 @@
 <script>
 
 const exportCSV = 'csv';
+const exportPDF = 'pdf';
+
+import { jsPDF } from "jspdf";
+import autoTable from 'jspdf-autotable';
 
     export default {
         components: {
@@ -78,7 +83,7 @@ const exportCSV = 'csv';
 
         data() {
             return {
-                exportTypes: [exportCSV]
+                exportTypes: [exportCSV, exportPDF]
             }
         },
 
@@ -97,42 +102,79 @@ const exportCSV = 'csv';
                 }
             },
 
-            goExport(type){
+            async goExport(type){
                 let fileContents = '';
-                let fileName = '';
+                let fileName = this.reportTitle;
                 let fileMimeType = '';
+                let download = true;
+
+                let simpleArr = [];
+
+                for (let i=0; i<this.items.length; i++){
+                    simpleArr.push([]);
+                    for (let j=0; j<this.headers.length; j++){
+                        let value = this.items[i];
+                        let digValues = this.headers[j].value.split(".");
+                        for (let k=0; k<digValues.length; k++){
+                            if (typeof(value) !== 'undefined'){
+                                value = value[digValues[k]];
+                            }
+                        }
+                        if (typeof(value) !== 'undefined'){
+                            simpleArr[i].push(value);
+                        }else{
+                            simpleArr[i].push('');
+                        }
+                    }
+                }
+
                 if (type === exportCSV){
                     fileMimeType = 'text/csv';
-                    fileName = this.reportTitle + ".csv";
+                    fileName += ".csv";
                     for (let i=0; i<this.headers.length; i++){
                         fileContents += '"' + this.headers[i].text + '",'
                     }
                     //replace last , with newline
                     fileContents = fileContents.slice(0,-1) + "\n";
 
-                    for (let i=0; i<this.items.length; i++){
-                        for (let j=0; j<this.headers.length; j++){
-                            let value = this.items[i];
-                            let digValues = this.headers[j].value.split(".");
-                            for (let k=0; k<digValues.length; k++){
-                                if (typeof(value) !== 'undefined'){
-                                    value = value[digValues[k]];
-                                }
-                            }
-                            if (typeof(value) !== 'undefined'){
-                                fileContents += '"' + value + '",';
-                            }else{
-                                fileContents += '"' + '",';
-                            }
+                    for (let i=0; i<simpleArr.length; i++){
+                        for (let j=0; j<simpleArr[i].length; j++){
+                            fileContents += '"' + simpleArr[i][j] + '",';
                         }
                         //replace last comma with newline
                         fileContents = fileContents.slice(0,-1) + "\n";
                     }
+                    
                     //remove last newline
-                    fileContents = fileContents.slice(0,-1)
+                    fileContents = fileContents.slice(0,-1);
+                }else if (type === exportPDF){
+                    fileName += ".pdf";
+                    download = false;
+                    let fileContents = new jsPDF({
+                        orientation: "landscape",
+                        unit: "in",
+                        format: [11, 8.5]
+                    });
+
+                    let headers = [];
+                    for (let i=0; i<this.headers.length; i++){
+                        headers.push(this.headers[i].text);
+                    }
+
+
+                    fileMimeType = "application/pdf";
+                    fileContents.text(this.reportTitle, 3, .5);
+                    autoTable(fileContents, {
+                        head: [headers],
+                        body: simpleArr,
+                        startY: 1,
+                    })
+                    fileContents.save(fileName);
                 }
-                let download = require('downloadjs');
-                download(fileContents, fileName, fileMimeType);
+                if (download){
+                    let download = require('downloadjs');
+                    download(fileContents, fileName, fileMimeType);
+                }
             }
         },