+++ frontend/src/i18n/en.js
@@ -292,6 +292,8 @@ export default {
         "Resource Notes": "Resource Notes",
 
         "Field Tags": "Field Tags",
+        "Supplemental Information": "Supplemental Information",
+        "Preview": "Preview",
 
         "help.dataset.name": "The name of the dataset",
         "help.dataset.description": "A description of the dataset",